Add channel presets, admin UI overhaul, and Live TV fixes for v0.0.2.3.
Ready-made Binarygeek119 channels with GitHub logo sync, decimal channel numbers, and corrected M3U/XMLTV public URLs.

+++ Jellyfin.Plugin.FinTV/Domain/Entities.cs
@@ -4,7 +4,7 @@ public class Channel
 {
     public Guid Id { get; set; } = Guid.NewGuid();
 
-    public int Number { get; set; }
+    public decimal Number { get; set; }
 
     public string Name { get; set; } = string.Empty;
